Added transformation (rotation) and map scrolling
The player now rotates towards mouse cursor, and map scrolls in the
direction where player is walking

+++ Windows MSVC10/Scrapheap Spaceships/OurMap.h
@@ -38,10 +38,10 @@ public:
 
 	// Setters and getters for map offset values
 
-	int getOffsetX();
-	int getOffsetY();
-	void setOffsetX(int offset);
-	void setOffsetY(int offset);
+	float getOffsetX();
+	float getOffsetY();
+	void setOffsetX(float offset);
+	void setOffsetY(float offset);
 
 	void draw();	// Draws the map
 private:
@@ -76,7 +76,7 @@ private:
 
 	TwoDimIntegerVector map; // This vector contains all the numbers which represent tiles
 
-	int xOffset, yOffset; // Where the camera is at the map
+	float xOffset, yOffset; // Where the camera is at the map
 
 	int mapColumns;
 	int mapSize;
